feat: Add valations and tests for spokenlanguages and insurance
